pyproject.toml
[tool.poetry]
name = "my_rag"
version = "0.1.0"
description = ""
authors = ["makenow <makenow90@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "3.11.5"
chromadb = "0.4.24"
notebook = "^7.2.1"
ipykernel = "^6.29.4"
langsmith = "^0.1.93"
pdfplumber = "^0.11.2"
langchain-teddynote = "0.0.28"
langchain = "^0.2.12"
langchain-community = "^0.2.11"
langchain-openai = "0.1.19"
langchain-core = "0.2.28"
langchain-huggingface = "^0.0.3"
numpy = "1.24"
faiss-cpu = "^1.8.0.post1"
pymupdf = "^1.24.10"
requests = "^2.32.3"
image = "^1.5.33"
markdownify = "^0.13.1"
python-dotenv = "^1.0.1"
accelerate = "^0.34.2"
transformers = "^4.44.2"
ninja = "^1.11.1.1"
torch = {version = "^2.4.1+cu121", source = "torch_cu121"}
torchvision = {version = "^0.19.1+cu121", source = "torch_cu121"}
torchaudio = {version = "^2.4.1+cu121", source = "torch_cu121"}
aio-pika = "^9.4.3"
discord-py = "^2.4.0"
pika = "^1.3.2"
konlpy = "^0.6.0"
jpype1 = "^1.5.0"

[[tool.poetry.source]]
name = "torch_cu121"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
